refactor(lance): specialize family implementation
Consolidate Lance decoder, vision, weights, and checkpoint staging.

Keep the canonical implementation entirely family-owned.

Remove unreachable generic builder copies.

Preserve image-understanding build, debug, and runtime behavior.
