+++ python/tensorrt_model_connect/families/lance/model/components/__init__.py
@@ -1,6 +1,4 @@
 # SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-from .plugin import plugin
-
-__all__ = ["plugin"]
+"""Family-owned specialized components."""
